Project wide usage of BigDecimal when dealing with money
This is the recommended approach when working with money. This also
fixes the issue where there was multiple trailing zeros in Receipt
totalValue, for example: 0.300000000000000004

+++ rvm/src/main/java/org/autumn24/items/Item.java
@@ -19,6 +19,7 @@ package org.autumn24.items;
 
 import org.autumn24.excpetion.InvalidItemSizeException;
 
+import java.math.BigDecimal;
 import java.util.Random;
 import java.util.stream.DoubleStream;
 
@@ -46,7 +47,7 @@ public interface Item {
 		determineItemValue();
 	}
 
-	double getDeterminedValue();
+	BigDecimal getDeterminedValue();
 
 	ItemMaterial getItemMaterial();
 }
